orumaiv/requirements.txt
# API e Framework Web
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
email-validator>=2.1.0

# Google AI SDK
google-generativeai>=0.5.0
google-adk>=0.3.0

# Observabilidade
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-jaeger>=1.25.0
prometheus-client>=0.19.0

# Banco de dados e cache
motor>=3.3.0  # Cliente MongoDB assíncrono
redis>=5.0.0
aioredis>=2.0.0

# Mensageria
aiokafka>=0.10.0

# Segurança
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
cryptography>=42.0.0

# Utilidades
httpx>=0.27.0
python-multipart>=0.0.7
tenacity>=8.2.0
ujson>=5.9.0 